Patch from Jeff Senn.  Use Py_ssize_t instead of int where applicable in the pickling code.
git-svn-id: http://svn.python.org/projects/stackless/branches/release25-maint@52026 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Stackless/pickling/prickelpit.c
@@ -1376,8 +1376,8 @@ static int init_methodtype(void)
 typedef struct {
 	PyObject_HEAD
 	PyDictObject *di_dict; /* Set to NULL when iterator is exhausted */
-	int di_used;
-	int di_pos;
+	Py_ssize_t di_used;
+	Py_ssize_t di_pos;
 	binaryfunc di_select;
 } dictiterobject;
 
@@ -1393,7 +1393,7 @@ static PyObject *
 dictiterkey_reduce(dictiterobject *di)
 {
     PyObject *tup, *list, *key;
-    int i;
+    Py_ssize_t i;
 
     list = PyList_New(0);
     if (list == NULL)
@@ -1433,7 +1433,7 @@ static PyObject *
 dictitervalue_reduce(dictiterobject *di)
 {
     PyObject *tup, *list, *value;
-    int i;
+    Py_ssize_t i;
 
     list = PyList_New(0);
     if (list == NULL)
@@ -1473,7 +1473,7 @@ static PyObject *
 dictiteritem_reduce(dictiterobject *di)
 {
     PyObject *tup, *list, *key, *value, *res;
-    int i;
+    Py_ssize_t i;
 
     list = PyList_New(0);
     if (list == NULL)
@@ -1579,7 +1579,7 @@ static PyObject *
 setiter_reduce(setiterobject *it)
 {
     PyObject *list, *set, *elem;
-    int i;
+    Py_ssize_t i;
 
     list = PyList_New(0);
     if (list == NULL)
